make point_color_cell
make point, color_ cell.c&h
make other fiies.
