+++ board.c
@@ -1,0 +1,1 @@
+#include "board.h"
