+++ main.c
@@ -1,9 +1,15 @@
-﻿#include <stdio.h>
-#include <stdint.h>
+﻿//#include <stdio.h>
+//#include <stdint.h>
+//#include <stdbool.h>
+
+#include <stdio.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <windows.h>
 
 #include "blocks.h"
+#include "cell.h"
+#include "color.h"
 #include "key_driver.h"
 
 #define _CRT_SECURE_NO_WARNINGS
@@ -14,56 +20,8 @@
 #define RESIZE_COLS (DEFAULT_COLS-30) //=90
 #define RESIZE_ROWS 30
 
-//github desktop test
-//추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-typedef enum _color_t {
-    BLACK, //0
-    BLUE, // 1
-    GREEN,
-    CYAN,
-    RED,
-    PURPLE,
-    YELLOW,
-    WHITE,
-    GRAY,
-    LIGHT_BLUE,
-    LIGHT_GREEN,
-    LIGHT_CYAN,
-    LIGHT_RED,
-    LIGHT_PURBPLE,
-    LIGHT_YELLOW,
-    LIGHT_WHITE,
-} color_t;
-
 
 //추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-char* color_strings[16] = {
-    "black", // 검정
-    "blue", // 파랑
-    "green", // 초록
-    "cyan", // 청록
-    "red", // 빨강
-    "purple", // 자주
-    "yellow", // 노랑
-    "white", // 흰색
-    "gray", // 회색
-    "light blue", // 밝은 파랑
-    "light green", // 밝은 회색
-    "light cyan", // 밝은 청록
-    "light red", // 밝은 빨강
-    "light purple", // 밝은 자주
-    "light yellow", // 밝은 노랑
-    "light white" // 밝은 흰색
-};
-
-char* get_color_string(color_t color) {
-    if (color >= BLACK && color <= LIGHT_WHITE) {
-        return color_strings[color]; // 유효한 색상 값이면 해당 문자열 반환
-    }
-    else {
-        return "unknown"; // 유효하지 않은 값이면 기본값 반환
-    }
-}
 
 typedef struct _cursor_info_t {
     uint32_t x;
@@ -201,16 +159,9 @@ void console_display_info(const windows_console_t* console) {
     printf("[%d] Cursor (x,y,hide) = %d,%d,%d)\r\n", ++dummy_count, console->cursor_info.x, console->cursor_info.y, console->cursor_info.is_hide);
 }
 
-typedef struct _point_t {
-    uint32_t x;
-    uint32_t y;
-} point_t;
 
-typedef struct _cell_t {
-    point_t point; // cell의 위치
-    color_t color; // cell의 색상
-    cell_attributes_t att; // cell의 속성
-} cell_t;
+
+
 
 void cell_init(cell_t* cell, int x, int y, cell_attributes_t att, color_t color) {
     cell->point.x = x;
