+++ point.h
@@ -1,0 +1,9 @@
+#ifndef __POINT_H__
+#define __POINT_H__
+
+typedef struct _point_t {
+    uint32_t x;
+    uint32_t y;
+} point_t;
+
+#endif
